requirements.txt
streamlit
PyPDF2
python-docx
pymupdf
spacy
en_core_web_sm
ollama
sentence-transformers
faiss-cpu
fpdf


